pyproject.toml
[project]
dependencies = ['PyYAML >= 6.0.1', 'python-dotenv >= 1.0.1', 'pandas == 2.2.2',
                'et-xmlfile == 1.1.0', 'openpyxl == 3.1.5']

name = "dependency_template" 
version = "1.0.2"
authors = [
  { name="Elias Attias", email="attiaselias@gmail.com" }
]
description = "Dependency Template"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent"
]

[tool.setuptools.package-data]
dependency_template = ["communication/*"]